requirements.txt
Flask==3.1.0
functions_framework==3.8.2
line_bot_sdk==3.14.2
openai==1.55.0
protobuf==5.28.3
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
PyYAML==6.0.2
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.21.1
google-auth==2.36.0
